Move `install_python()` to `install` module
